+++ docs/src/pages/style/css-baseline/css-baseline.md
@@ -1,19 +1,19 @@
 ---
-components: Reboot
+components: CssBaseline
 ---
 
-# Reboot
+# CssBaseline
 
 You might be familiar with [normalize.css](https://github.com/necolas/normalize.css), a collection of HTML element and attribute style-normalizations.
-Material-UI provides a `Reboot` component to kickstart an elegant, consistent, and simple baseline to build upon.
+Material-UI provides a `CssBaseline` component to kickstart an elegant, consistent, and simple baseline to build upon.
 
 ```jsx
-import Reboot from 'material-ui/Reboot';
+import CssBaseline from 'material-ui/CssBaseline';
 
 function MyApp() {
   return (
     <div>
-      <Reboot />
+      <CssBaseline />
       {/* The rest of your application */}
     </div>
   );
@@ -33,7 +33,7 @@ The `<html>` and `<body>` elements are updated to provide better page-wide defau
 ### Layout
 
 - `box-sizing` is set globally on the `<html>` element to `border-box`.
-Every element—including `*::before` and `*::after` are declared to inherit this property, 
+Every element—including `*::before` and `*::after` are declared to inherit this property,
 which ensures that the declared width of the element is never exceeded due to padding or border.
 
 ### Typography
