+++ pages/api/css-baseline.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import withRoot from 'docs/src/modules/components/withRoot';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from './reboot.md';
+import markdown from './css-baseline.md';
 
 function Page() {
   return <MarkdownDocs markdown={markdown} />;
