+++ pages/style/css-baseline.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import withRoot from 'docs/src/modules/components/withRoot';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from 'docs/src/pages/style/reboot/reboot.md';
+import markdown from 'docs/src/pages/style/css-baseline/css-baseline.md';
 
 function Page() {
   return <MarkdownDocs markdown={markdown} />;
