+++ src/CssBaseline/CssBaseline.d.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export interface CssBaselineProps {
+  children?: React.ReactElement<any>;
+}
+
+declare const CssBaseline: React.ComponentType<CssBaselineProps>;
+
+export default CssBaseline;
