+++ src/CssBaseline/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './CssBaseline';
+export * from './CssBaseline';
