--- src/Reboot/Reboot.d.ts
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-export interface RebootProps {
-  children?: React.ReactElement<any>;
-}
-
-declare const Reboot: React.ComponentType<RebootProps>;
-
-export default Reboot;
