--- src/Reboot/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from './Reboot';
-export * from './Reboot';
