+++ src/index.d.ts
@@ -63,6 +63,7 @@ export { default as Card, CardActions, CardContent, CardHeader, CardMedia } from
 export { default as Checkbox } from './Checkbox';
 export { default as Chip } from './Chip';
 export { default as ClickAwayListener } from './utils/ClickAwayListener';
+export { default as CssBaseline } from './CssBaseline';
 export {
   default as Dialog,
   DialogActions,
@@ -102,7 +103,6 @@ export { default as Popover } from './Popover';
 export { default as Portal } from './Portal';
 export { CircularProgress, LinearProgress } from './Progress';
 export { default as Radio, RadioGroup } from './Radio';
-export { default as Reboot } from './Reboot';
 export { default as Select } from './Select';
 export { default as Snackbar, SnackbarContent } from './Snackbar';
 export { default as Stepper, Step, StepButton, StepContent, StepIcon, StepLabel } from './Stepper';
